# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/platform/accounts/{id}/status": {
      "patch": {
        "operationId": "platformAccounts_setStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "PlatformAccounts"
        ]
      }
    },
...
    "/api/v1/tenant/accounts/{id}/status": {
      "patch": {
        "operationId": "tenantAccounts_setStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "TenantAccounts"
        ]
      }
    },
    "/api/v1/tenant/accounts/{id}/roles": {
      "patch": {
        "operationId": "tenantAccounts_setRoles",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "TenantAccounts"
        ]
      }
    },
...
    "/api/v1/tenant/customers/{id}/account": {
      "get": {
        "operationId": "customers_account",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Customers"
        ]
      },
      "post": {
        "operationId": "customers_bind",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Customers"
        ]
      }
    },
...
    "/api/v1/tenant/settlements/{id}": {
      "get": {
        "operationId": "settlements_detail",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Settlements"
        ]
      }
    },
    "/api/v1/tenant/settlements/{id}/items": {
      "post": {
        "operationId": "settlements_addItems",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Settlements"
        ]
      }
    },
    "/api/v1/tenant/settlements/{id}/review": {
      "post": {
        "operationId": "settlements_review",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Settlements"
        ]
      }
    },
    "/api/v1/tenant/settlements/{id}/approve": {
      "post": {
        "operationId": "settlements_approve",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Settlements"
        ]
      }
    },
    "/api/v1/tenant/settlements/{id}/pay": {
      "post": {
        "operationId": "settlements_pay",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Settlements"
        ]
      }
    },
    "/api/v1/tenant/settlements/{id}/void": {
      "post": {
        "operationId": "settlements_voidBatch",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Settlements"
        ]
      }
    },
...
    "/api/v1/tenant/players/{id}/account": {
      "get": {
        "operationId": "players_account",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Players"
        ]
      },
      "post": {
        "operationId": "players_bind",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Players"
        ]
      }
    },
...
    "/api/v1/platform/tenants/{tenantId}/package": {
      "post": {
        "operationId": "platformBilling_assign",
        "parameters": [
          {
            "name": "tenantId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "PlatformBilling"
        ]
      }
    },
    "/api/v1/platform/tenants/{tenantId}/activate": {
      "post": {
        "operationId": "platformBilling_activate",
        "parameters": [
          {
            "name": "tenantId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "PlatformBilling"
        ]
      }
    },
    "/api/v1/platform/subscriptions/{subscriptionId}/renew": {
      "post": {
        "operationId": "platformBilling_renew",
        "parameters": [
          {
            "name": "subscriptionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "PlatformBilling"
        ]
      }
    },
...
  },
  "info": {
    "title": "陪玩门店多租户 SaaS API",
    "description": "一期 API。金额字段一律为十进制字符串分（MoneyFen），禁止 number/浮点。",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "tenant/platform access token"
      }
    },
...
  }
}